Refactor directory package as a standard go package practice
